Add missing logic to data migration, cleanup and comment

+++ tests/conftest.py
@@ -539,7 +539,9 @@ def db(app, request):
         connection = app.db.engine.connect()
         connection.begin()
         for table in reversed(models.Base.metadata.sorted_tables):
-            connection.execute(table.delete())
+            # Delete table only if it exists
+            if app.db.engine.dialect.has_table(connection, table):
+                connection.execute(table.delete())
         connection.close()
 
     request.addfinalizer(drop_all)
